+++ src/main/java/team_questio/questio/gpt/service/util/dto/response/GPTChoice.java
@@ -1,0 +1,8 @@
+package team_questio.questio.gpt.service.util.dto.response;
+
+import team_questio.questio.gpt.service.util.dto.GPTMessage;
+
+public record GPTChoice(
+        GPTMessage message
+) {
+}
