--- src/main/java/team_questio/questio/portfolio/service/util/dto/response/GPTChoice.java
@@ -1,8 +1,0 @@
-package team_questio.questio.portfolio.service.util.dto.response;
-
-import team_questio.questio.portfolio.service.util.dto.GPTMessage;
-
-public record GPTChoice(
-        GPTMessage message
-) {
-}
